# workorder-system/openapi.reference.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contoso Maintenance Actions",
    "description": "Endpoints on the Contoso Work Order & Warranty System (Azure App Service) that let the Copilot Studio agent check warranty status, predict maintenance risk with the Azure AI Foundry agent, and create work orders.",
    "version": "1.1.0"
  },
  "host": "app-contosowo-mvjskqas3y4lo.azurewebsites.net",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/checkWarranty": {
      "get": {
        "operationId": "checkWarranty",
        "summary": "Check equipment warranty status",
        "description": "Checks the warranty status of a piece of equipment by its asset ID (e.g., CE-OSC-1200). Returns whether it is under warranty and the number of days remaining.",
        "parameters": [
          {
            "name": "assetId",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "The asset ID of the equipment to check (e.g., CE-OSC-1200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Warranty status for the requested asset.",
            "schema": {
              "$ref": "#/definitions/Warranty"
            },
            "examples": {
              "application/json": {
                "assetId": "CE-OSC-1200",
                "name": "Digital Storage Oscilloscope",
                "model": "Contoso WaveView DSO-4000X",
                "warrantyProvider": "Contoso Instruments Division",
                "installDate": "2023-03-14",
                "warrantyExpiry": "2027-03-14",
                "underWarranty": true,
                "status": "Active",
                "daysRemaining": 249,
                "supportContact": "test.equipment.support@contoso.com",
                "checkedAt": "2026-07-09T12:00:00.000Z"
              }
            }
          },
          "400": {
            "description": "Missing required assetId query parameter.",
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "examples": {
              "application/json": {
                "error": "The \"assetId\" query parameter is required."
              }
            }
          },
          "404": {
            "description": "Equipment not found.",
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "examples": {
              "application/json": {
                "error": "Equipment 'CE-XYZ-0000' not found."
              }
            }
          }
        }
      }
    },
    "/foundry/predict": {
      "post": {
        "operationId": "predictMaintenance",
        "summary": "Predict maintenance risk for equipment",
        "description": "Uses the Azure AI Foundry 'Predictive Maintenance Insights' agent to reason over an asset's warranty status and recent work-order history and return a risk score (0-100), a risk level, and a recommended maintenance action.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PredictionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Risk assessment and recommended action for the requested asset.",
            "schema": {
              "$ref": "#/definitions/Prediction"
            },
            "examples": {
              "application/json": {
                "assetId": "CE-LAS-3300",
                "equipmentName": "CO2 Laser Cutter",
                "model": "Contoso PrecisionCut LC-150",
                "location": "Enclosure Fabrication Shop - Bay 2",
                "warrantyStatus": "Expired",
                "underWarranty": false,
                "riskScore": 78,
                "riskLevel": "Critical",
                "confidence": 0.82,
                "recommendedAction": "Take CE-LAS-3300 out of production for a laser tube power measurement and full optical path alignment within the week.",
                "recommendedPriority": "Critical",
                "recommendedWithinDays": 5,
                "rationale": "Two work orders in the last 90 days against a 78-day mean interval shows the failure cadence is accelerating.",
                "suggestedWorkOrderTitle": "Laser tube power check and optical realignment",
                "source": "azure-ai-foundry",
                "generatedAt": "2026-08-18T10:00:00.000Z"
              }
            }
          },
          "400": {
            "description": "Missing required assetId field.",
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "examples": {
              "application/json": {
                "error": "The \"assetId\" field is required."
              }
            }
          },
          "404": {
            "description": "Equipment not found.",
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "examples": {
              "application/json": {
                "error": "Equipment 'CE-XYZ-0000' not found."
              }
            }
          },
          "502": {
            "description": "The Azure AI Foundry agent could not be reached and fallback is disabled.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/createWorkOrder": {
      "post": {
        "operationId": "createWorkOrder",
        "summary": "Create a maintenance work order",
        "description": "Creates a maintenance work order for a piece of equipment. Requires the asset ID and a short title; accepts priority, description, and requestedBy.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WorkOrderRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The created work order.",
            "schema": {
              "$ref": "#/definitions/WorkOrder"
            },
            "examples": {
              "application/json": {
                "id": "WO-2026-0007",
                "assetId": "CE-LAS-3300",
                "equipmentName": "CO2 Laser Cutter",
                "equipmentModel": "LC-1390 Pro",
                "location": "Fabrication Bay 2",
                "title": "Laser cutter needs service",
                "description": "Beam alignment drifting; schedule maintenance.",
                "priority": "High",
                "status": "Open",
                "requestedBy": "Copilot Agent",
                "assignedTo": null,
                "underWarrantyAtCreation": false,
                "createdAt": "2026-07-09T12:00:00.000Z",
                "updatedAt": "2026-07-09T12:00:00.000Z",
                "notes": []
              }
            }
          },
          "400": {
            "description": "Validation failed (missing assetId or title, or invalid priority).",
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "examples": {
              "application/json": {
                "error": "The \"assetId\" and \"title\" fields are required."
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "PredictionRequest": {
      "type": "object",
      "required": [
        "assetId"
      ],
      "properties": {
        "assetId": {
          "type": "string",
          "description": "The asset ID of the equipment to assess (e.g., CE-LAS-3300).",
          "example": "CE-LAS-3300"
        },
        "historyLimit": {
          "type": "integer",
          "format": "int32",
          "description": "How many recent work orders to reason over (1-20, default 5).",
          "example": 5
        },
        "workOrderHistory": {
          "type": "array",
          "description": "Optional extra history the caller observed that is not yet ticketed in the system.",
          "items": {
            "$ref": "#/definitions/WorkOrderHistoryItem"
          }
        }
      }
    },
    "WorkOrderHistoryItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-nullable": true,
          "example": "WO-2026-0009"
        },
        "title": {
          "type": "string",
          "example": "Incomplete cuts on 6 mm acrylic"
        },
        "description": {
          "type": "string",
          "example": "Operator reports beam not cutting through."
        },
        "priority": {
          "type": "string",
          "example": "High"
        },
        "status": {
          "type": "string",
          "example": "Open"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "x-nullable": true,
          "example": "2026-08-07T09:12:00.000Z"
        }
      }
    },
    "RiskFactor": {
      "type": "object",
      "properties": {
        "factor": {
          "type": "string",
          "example": "Accelerating failure cadence"
        },
        "impact": {
          "type": "string",
          "enum": [
            "Low",
            "Medium",
            "High"
          ],
          "example": "High"
        },
        "detail": {
          "type": "string",
          "example": "2 work orders in the last 90 days versus a 78-day mean interval."
        }
      }
    },
    "PredictionSignals": {
      "type": "object",
      "properties": {
        "ageYears": {
          "type": "number",
          "format": "float",
          "x-nullable": true,
          "example": 4
        },
        "totalWorkOrders": {
          "type": "integer",
          "format": "int32",
          "example": 4
        },
        "openWorkOrders": {
          "type": "integer",
          "format": "int32",
          "example": 2
        },
        "highOrCriticalWorkOrders": {
          "type": "integer",
          "format": "int32",
          "example": 3
        },
        "workOrdersLast90Days": {
          "type": "integer",
          "format": "int32",
          "example": 2
        },
        "workOrdersLast365Days": {
          "type": "integer",
          "format": "int32",
          "example": 3
        },
        "daysSinceLastWorkOrder": {
          "type": "integer",
          "format": "int32",
          "x-nullable": true,
          "example": 11
        },
        "meanDaysBetweenWorkOrders": {
          "type": "integer",
          "format": "int32",
          "x-nullable": true,
          "example": 78
        }
      }
    },
    "Prediction": {
      "type": "object",
      "properties": {
        "assetId": {
          "type": "string",
          "example": "CE-LAS-3300"
        },
        "equipmentName": {
          "type": "string",
          "example": "CO2 Laser Cutter"
        },
        "model": {
          "type": "string",
          "example": "Contoso PrecisionCut LC-150"
        },
        "location": {
          "type": "string",
          "example": "Enclosure Fabrication Shop - Bay 2"
        },
        "warrantyStatus": {
          "type": "string",
          "example": "Expired"
        },
        "underWarranty": {
          "type": "boolean",
          "example": false
        },
        "riskScore": {
          "type": "integer",
          "format": "int32",
          "description": "Failure risk from 0 (low) to 100 (imminent).",
          "example": 78
        },
        "riskLevel": {
          "type": "string",
          "enum": [
            "Low",
            "Moderate",
            "High",
            "Critical"
          ],
          "example": "Critical"
        },
        "confidence": {
          "type": "number",
          "format": "float",
          "example": 0.82
        },
        "recommendedAction": {
          "type": "string",
          "example": "Take CE-LAS-3300 out of production for a laser tube power measurement within the week."
        },
        "recommendedPriority": {
          "type": "string",
          "enum": [
            "Low",
            "Medium",
            "High",
            "Critical"
          ],
          "example": "Critical"
        },
        "recommendedWithinDays": {
          "type": "integer",
          "format": "int32",
          "example": 5
        },
        "rationale": {
          "type": "string",
          "example": "Two work orders in the last 90 days against a 78-day mean interval shows the failure cadence is accelerating."
        },
        "suggestedWorkOrderTitle": {
          "type": "string",
          "example": "Laser tube power check and optical realignment"
        },
        "riskFactors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RiskFactor"
          }
        },
        "signals": {
          "$ref": "#/definitions/PredictionSignals"
        },
        "workOrderHistory": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkOrderHistoryItem"
          }
        },
        "source": {
          "type": "string",
          "description": "'azure-ai-foundry' when the Foundry agent answered, 'local-heuristic' when the built-in fallback did.",
          "enum": [
            "azure-ai-foundry",
            "local-heuristic"
          ],
          "example": "azure-ai-foundry"
        },
        "foundryAgentName": {
          "type": "string",
          "x-nullable": true,
          "example": "predictive-maintenance-insights"
        },
        "fallbackReason": {
          "type": "string",
          "x-nullable": true,
          "example": null
        },
        "generatedAt": {
          "type": "string",
          "format": "date-time",
          "example": "2026-08-18T10:00:00.000Z"
        }
      }
    },
    "Warranty": {
      "type": "object",
      "properties": {
        "assetId": {
          "type": "string",
          "example": "CE-OSC-1200"
        },
        "name": {
          "type": "string",
          "example": "Digital Storage Oscilloscope"
        },
        "model": {
          "type": "string",
          "example": "Contoso WaveView DSO-4000X"
        },
        "warrantyProvider": {
          "type": "string",
          "example": "Contoso Instruments Division"
        },
        "installDate": {
          "type": "string",
          "example": "2023-03-14"
        },
        "warrantyExpiry": {
          "type": "string",
          "example": "2027-03-14"
        },
        "underWarranty": {
          "type": "boolean",
          "example": true
        },
        "status": {
          "type": "string",
          "example": "Active"
        },
        "daysRemaining": {
          "type": "integer",
          "format": "int32",
          "example": 249
        },
        "supportContact": {
          "type": "string",
          "example": "test.equipment.support@contoso.com"
        },
        "checkedAt": {
          "type": "string",
          "example": "2026-07-09T12:00:00.000Z"
        }
      }
    },
    "WorkOrderRequest": {
      "type": "object",
      "required": [
        "assetId",
        "title"
      ],
      "properties": {
        "assetId": {
          "type": "string",
          "description": "The asset ID of the equipment (e.g., CE-LAS-3300).",
          "example": "CE-LAS-3300"
        },
        "title": {
          "type": "string",
          "description": "Short summary of the work to be done.",
          "example": "Laser cutter needs service"
        },
        "priority": {
          "type": "string",
          "description": "Work order priority.",
          "enum": [
            "Low",
            "Medium",
            "High",
            "Critical"
          ],
          "example": "High"
        },
        "description": {
          "type": "string",
          "description": "Detailed description of the issue.",
          "example": "Beam alignment drifting; schedule maintenance."
        },
        "requestedBy": {
          "type": "string",
          "description": "Name of the person or system requesting the work order.",
          "example": "Copilot Agent"
        }
      }
    },
    "WorkOrder": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "WO-2026-0007"
        },
        "assetId": {
          "type": "string",
          "example": "CE-LAS-3300"
        },
        "equipmentName": {
          "type": "string",
          "example": "CO2 Laser Cutter"
        },
        "equipmentModel": {
          "type": "string",
          "example": "LC-1390 Pro"
        },
        "location": {
          "type": "string",
          "example": "Fabrication Bay 2"
        },
        "title": {
          "type": "string",
          "example": "Laser cutter needs service"
        },
        "description": {
          "type": "string",
          "example": "Beam alignment drifting; schedule maintenance."
        },
        "priority": {
          "type": "string",
          "example": "High"
        },
        "status": {
          "type": "string",
          "example": "Open"
        },
        "requestedBy": {
          "type": "string",
          "example": "Copilot Agent"
        },
        "assignedTo": {
          "type": "string",
          "x-nullable": true,
          "example": null
        },
        "underWarrantyAtCreation": {
          "type": "boolean",
          "example": false
        },
        "createdAt": {
          "type": "string",
          "example": "2026-07-09T12:00:00.000Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2026-07-09T12:00:00.000Z"
        },
        "notes": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "The \"assetId\" query parameter is required."
        }
      }
    }
  }
}